+++ tests/block_alignment.rs
@@ -8,7 +8,7 @@ fn center_alignment() {
     let window = Size::new(500.0, 500.0);
 
     let child = EmptyLayout::new().intrinsic_size(IntrinsicSize::fixed(250.0, 350.0));
-    let mut root = BlockLayout::new(Box::new(child))
+    let mut root = BlockLayout::new(child)
         .main_axis_alignment(AxisAlignment::Center)
         .cross_axis_alignment(AxisAlignment::Center)
         .padding(Padding::all(24.0))
@@ -29,7 +29,7 @@ fn test_start_alignment() {
     let padding = 32;
 
     let child_1 = EmptyLayout::new().intrinsic_size(IntrinsicSize::fixed(240.0, 40.0));
-    let mut root = BlockLayout::new(Box::new(child_1)).padding(Padding::all(32.0));
+    let mut root = BlockLayout::new(child_1).padding(Padding::all(32.0));
     root.set_position(Position::new(20.0, 500.0));
 
     solve_layout(&mut root, window);
@@ -47,7 +47,7 @@ fn test_end_alignment() {
     let padding = 32;
 
     let child_1 = EmptyLayout::new().intrinsic_size(IntrinsicSize::fixed(240.0, 40.0));
-    let mut root = BlockLayout::new(Box::new(child_1))
+    let mut root = BlockLayout::new(child_1)
         .padding(Padding::all(32.0))
         .main_axis_alignment(AxisAlignment::End)
         .cross_axis_alignment(AxisAlignment::End);