+++ CHANGELOG.md
@@ -5,3 +5,4 @@ All notable changes to agape will be documented in this file.
 ## 0.2.0 (unreleased)
 
 - Make all layout fields private.
+- `BlockLayout::new` now takes in a `Layout` instead of a `Box<dyn Layout>`;